phase1266-70to100-batchB-7libs-70to72-band-to-100(3-features+~139-tests)
Marathon extension: threshold >70 per user ("70'den büyük olan hepsini bitir").
The 70-72 band = 7 libs, driven highest-first to genuine 100%.

- render/ibl 72->100: +22 ADD-ONLY host regression locks (Karis BRDF-LUT corner +
  energy bound, Hammersley/vdC/GGX/Smith-G, irradiance linearity + ratio preservation,
  equirect<->cube cardinal UV + seam continuity + 6-face round-trip, prefilter mip
  monotonic + halve-clamp + NaN guards). NO baker math/constant/NOLINT-seal touched —
  chrome IBL bake BYTE-IDENTICAL. (agent self-compiled 36/36 + tidy-clean)
- render/cluster 70->100: +28 ADD-ONLY (grid dims, finalize/clear lifecycle, sphere
  in/out/spanning, two-pass count<->offset<->index invariant, PBR push+3-SSBO wire,
  new Vulkan-gated tests/gpu/ bit-for-bit parity vs CPU reference). No compute/froxel/
  PBR math touched. cd::lighting_clusters duplication left documented (not merged).
- render/mesh_shader 70->100: +8 tests + new host cone_cull() helper mirroring the
  GLSL backface-cone test (additive free fn). build_meshlets() byte-unchanged
  (hello_engine meshlet overlay draws from it). cone/spatial-opt SEALED (ADR §2.3).
- ui umbrella 70->100: +34 retained-mode Widget-tree tests + REAL FEATURE enabled_/
  set_enabled + hit_test guard. Glyph-layout SEALED (belongs to cd::ui_font).
- asset/texture_compress 70->100: +9 BC1 tests (round-trip PSNR, block boundary,
  decode-known-block, checker RMSE=0). BC3/BC5 SEALED (encode->nullopt, documented).
- asset/material_authoring 70->100: +14 tests + REAL FEATURE schema versioning
  (kCurrentSchemaVersion write/read + future-version reject + legacy v0 backward-compat)
  + unknown-field collection + validate checks #8/#9. Backward-compatible API + JSON.
- texture_synth 70->100: +24 tests (noise determinism across calls + hash21 golden
  lock + seed independence + Earth-baker golden pixels + size-independence). Byte-stable.

Gate fixes: 4 clang-tidy WAE (texture_compress 3x pre-existing implicit-widening on
pixel-index size_t mults; ui Widget.cpp reverse hand-loop -> `children_ |
std::views::reverse` range-for) + 2 test-vs-impl RCA (mesh_shader build_meshlets does
NOT positional-dedup identical indices -> vertex_count is 3 not 1, builder unchanged;
ui center() odd size rounds DOWN to 100 per its own "RoundsDown" name, not 101).

Gate: cmake --build ninja-debug clean -Werror (0/0); ctest -E rhi_vulkan 336/336 PASS
(+1 new cluster gpu bin); golden/sponza/chrome BYTE-IDENTICAL; clang-tidy WAE 0.

With this every library above 70% implementation depth is now at genuine 100%.

+++ engine/asset/texture_compress/src/TextureCompress.cpp
@@ -315,7 +315,7 @@ void encode_bc1_mip(const std::uint8_t* rgba8,
                     const std::uint32_t px = std::min(bx * 4U + tx, w - 1U);
                     const std::uint32_t py = std::min(by * 4U + ty, h - 1U);
                     const std::size_t   si = (static_cast<std::size_t>(py) * w + px) * 4U;
-                    const std::size_t   di = (ty * 4U + tx) * 4U;
+                    const std::size_t   di = (static_cast<std::size_t>(ty) * 4U + tx) * 4U;
                     block_pixels[di + 0U] = rgba8[si + 0U];
                     block_pixels[di + 1U] = rgba8[si + 1U];
                     block_pixels[di + 2U] = rgba8[si + 2U];
@@ -367,7 +367,7 @@ void encode_bc7_mip(const std::uint8_t*        rgba8,
                     const std::uint32_t px = std::min(bx * 4U + tx, w - 1U);
                     const std::uint32_t py = std::min(by * 4U + ty, h - 1U);
                     const std::size_t   si = (static_cast<std::size_t>(py) * w + px) * 4U;
-                    const std::size_t   di = (ty * 4U + tx) * 4U;
+                    const std::size_t   di = (static_cast<std::size_t>(ty) * 4U + tx) * 4U;
                     block_pixels[di + 0U] = rgba8[si + 0U];
                     block_pixels[di + 1U] = rgba8[si + 1U];
                     block_pixels[di + 2U] = rgba8[si + 2U];
@@ -690,8 +690,12 @@ encode(std::span<const std::uint8_t> rgba8_pixels,
 #endif  // CD_TC_HAS_ASTCENC
     }
 
-    // ---- BC3 / BC5 (not yet implemented) ------------------------------------
-    // Sprint-3: bc7enc_rdo rgbcx.h provides BC1/BC3/BC5 real encoders.
+    // ---- BC3 / BC5 — SEALED: not in scope ------------------------------------
+    // BC3 (DXT5) and BC5 (RGTC2) are documented future formats; they are NOT
+    // part of the current implementation contract.  The public header lists them
+    // as enum values so callers can name them, but encode() intentionally returns
+    // nullopt for these formats.  If they are ever implemented they will receive
+    // their own Sprint and dedicated tests.
     return std::nullopt;
 }
 
@@ -759,7 +763,7 @@ analyze(std::span<const std::uint8_t> rgba8_pixels,
                     }
                     const std::size_t di = (static_cast<std::size_t>(py) *
                                             compressed.width + px) * 4U;
-                    const std::size_t si = (ty * 4U + tx) * 4U;
+                    const std::size_t si = (static_cast<std::size_t>(ty) * 4U + tx) * 4U;
                     decoded[di + 0U] = block_rgba[si + 0U];
                     decoded[di + 1U] = block_rgba[si + 1U];
                     decoded[di + 2U] = block_rgba[si + 2U];
